DataTypesNested: ignore paths with leading dot
The paths with leading dot can represent a full
path from JSON or schema-based protocols, for
example with the following row:

```
  {"SearchPhrase":"","count()":"8267016"}
```

The `SELECT .SearchPhrase` should be interpreted
as a full path to field `SearchPhrase`, not a
nested column. This becomes problematic when
extracting fields from nested structures
such as:

```
  {“Category”:{“Subcategory”:{“Field”:1}}
```

Where `Category.Subcategory.Field` would be
otherwise treated as an invalid `Nested` type field.